Applied MVC in my work
